Fix issue #109: simple formula write not working
This is based on the observation that for cells of number and formula types, the "t" attribute of a cell is optional for OpenOffice, and is actually harmful for MS Excel 2007 to interpret formula cells. Therefore once we remove that when cell's data type is number or formula, the issue is resolved. I have tested the code provided in the issue post, and verified that the generated excel file is OK both for OpenOffice and MS Excel.